refactor: migrate useAuth hook imports to auth-context for improved performance and state sharing
