+++ src/app/(auth)/register/page.tsx
@@ -20,7 +20,7 @@ import {
   User,
   Lock,
 } from "lucide-react";
-import { useAuth } from "@/hooks/use-auth";
+import { useAuth } from "@/contexts/auth-context";
 
 export default function RegisterPage() {
   const [name, setName] = useState("");
