+++ src/app/(public)/admin/page.tsx
@@ -4,7 +4,7 @@ import { AdminDashboard } from "@/components/admin";
 import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Skeleton } from "@/components/ui/skeleton";
 import { Shield } from "lucide-react";
-import { useAuth } from "@/hooks/use-auth";
+import { useAuth } from "@/contexts/auth-context";
 import { getUserRole } from "@/lib/user-utils";
 
 export default function AdminPage() {
