+++ src/components/auth-modal.tsx
@@ -26,7 +26,7 @@ import {
   AlertCircle,
   Loader2,
 } from "lucide-react";
-import { useAuth } from "@/hooks/use-auth";
+import { useAuth } from "@/contexts/auth-context";
 import { useAsync } from "@/hooks/use-api";
 import { authService } from "@/services/auth";
 import { toast } from "sonner";
