+++ src/components/library/library-action-button.tsx
@@ -24,7 +24,7 @@ import {
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { toast } from "sonner";
-import { useAuth } from "@/hooks/use-auth";
+import { useAuth } from "@/contexts/auth-context";
 import {
   useNovelLibraryStatus,
   useUpdateLibraryStatus,
