+++ src/components/profile/profile-settings.tsx
@@ -21,7 +21,7 @@ import {
   Trash2,
   Shield,
 } from "lucide-react";
-import { useAuth } from "@/hooks/use-auth";
+import { useAuth } from "@/contexts/auth-context";
 import { useAsync } from "@/hooks/use-api";
 import { authService } from "@/services/auth";
 import { toast } from "sonner";
